Add modes to our Babel preset (1.x)
