+++ packages/babel-preset-react-app/package.json
@@ -8,7 +8,11 @@
     "url": "https://github.com/facebookincubator/create-react-app/issues"
   },
   "files": [
-    "index.js"
+    "index.js",
+    "create.js",
+    "dev.js",
+    "prod.js",
+    "test.js"
   ],
   "dependencies": {
     "babel-plugin-dynamic-import-node": "1.1.0",
